chore: move the keep alive package to the session package
